requirements.txt
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==2.2.19
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uritemplate==4.1.1
urllib3==2.0.7
